fix: change the deduplication strategy
Since the next_id of both the head and tail of the circular queue is None (included in the finally returned json), the data deduplication strategy has been changed.
In addition, the potential error in using all_ids for deduplication in the previous submission has been fixed.
About this error, for example, during the crawling process, dynamic updates of previous data may lead to the addition or deletion of some data, thus causing dynamic changes in next_id. The strategy of judging next_id in all_ids may have a delay, that is, duplicate data has already been crawled, but due to the misalignment between the original data and the updated data, the crawling will stop at a certain moment, which has a certain coincidental nature.